Add delete message functionality for Twitter

+++ app/src/main/java/com/channels/androidsms/SmsMessage.java
@@ -55,7 +55,7 @@ public class SmsMessage extends AppCompatActivity implements Message {
      * Delete a message from content://sms/{id}, where id is set from the 2 param constructor
      */
     @Override
-    public Boolean delete(int id) {
+    public Boolean delete(long id) {
 
         try {
             context.getContentResolver().delete(
